+++ lizard_neerslagradar/testsettings.py
@@ -1,59 +1,43 @@
 import os
 
 from lizard_ui.settingshelper import setup_logging
-from lizard_ui.settingshelper import STATICFILES_FINDERS
+
+SETTINGS_DIR = os.path.dirname(os.path.realpath(__file__))
+BUILDOUT_DIR = os.path.abspath(os.path.join(SETTINGS_DIR, '..'))
+
+LOGGING = setup_logging(BUILDOUT_DIR)
 
 DEBUG = True
 TEMPLATE_DEBUG = True
-
-# SETTINGS_DIR allows media paths and so to be relative to this settings file
-# instead of hardcoded to c:\only\on\my\computer.
-SETTINGS_DIR = os.path.dirname(os.path.realpath(__file__))
-
-# BUILDOUT_DIR is for access to the "surrounding" buildout, for instance for
-# BUILDOUT_DIR/var/static files to give django-staticfiles a proper place
-# to place all collected static files.
-BUILDOUT_DIR = os.path.abspath(os.path.join(SETTINGS_DIR, '..'))
-LOGGING = setup_logging(BUILDOUT_DIR)
-
-# ENGINE: 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
-# In case of geodatabase, prepend with:
-# django.contrib.gis.db.backends.(postgis)
 DATABASES = {
-    # If you want to use another database, consider putting the database
-    # settings in localsettings.py. Otherwise, if you change the settings in
-    # the current file and commit them to the repository, other developers will
-    # also use these settings whether they have that database or not.
-    # One of those other developers is Jenkins, our continuous integration
-    # solution. Jenkins can only run the tests of the current application when
-    # the specified database exists. When the tests cannot run, Jenkins sees
-    # that as an error.
     'default': {
-        'NAME': os.path.join(BUILDOUT_DIR, 'var', 'sqlite', 'test.db'),
         'ENGINE': 'django.contrib.gis.db.backends.spatialite',
-        'USER': '',
-        'PASSWORD': '',
-        'HOST': '',  # empty string for localhost.
-        'PORT': '',  # empty string for default.
-        }
-    }
+        'NAME': os.path.join(BUILDOUT_DIR, 'test.db')
+    },
+}
+
 SITE_ID = 1
-SECRET_KEY = 'This is not secret but that is ok.'
+TIME_ZONE = 'Europe/Amsterdam'
+LANGUAGE_CODE = 'nl'
+LANGUAGES = (
+    ('nl', 'Nederlands'),
+    ('en', 'English'),
+)
 INSTALLED_APPS = [
     'lizard_neerslagradar',
+    'lizard_map',
     'lizard_ui',
-    'staticfiles',
+    'lizard_security',
     'compressor',
     'south',
     'django_nose',
-    'lizard_security',
-    'django_extensions',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
-    'django.contrib.sessions',
     'django.contrib.gis',
+    'django.contrib.sessions',
     'django.contrib.sites',
+    'django.contrib.staticfiles',
     ]
 ROOT_URLCONF = 'lizard_neerslagradar.urls'
 
@@ -70,16 +54,32 @@ MIDDLEWARE_CLASSES = (
 
 TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
 
-# Used for django-staticfiles (and for media files
-STATIC_URL = '/static_media/'
-MEDIA_URL = '/media/'
+# Used for django-staticfiles
+STATIC_URL = '/static/'
 STATIC_ROOT = os.path.join(BUILDOUT_DIR, 'var', 'static')
-MEDIA_ROOT = os.path.join(BUILDOUT_DIR, 'var', 'media')
-STATICFILES_FINDERS = STATICFILES_FINDERS
+STATICFILES_FINDERS = (
+    'django.contrib.staticfiles.finders.FileSystemFinder',
+    'django.contrib.staticfiles.finders.AppDirectoriesFinder',
+    'compressor.finders.CompressorFinder',
+)
 
+TEMPLATE_CONTEXT_PROCESSORS = (
+    # Default items.
+    "django.contrib.auth.context_processors.auth",
+    "django.core.context_processors.debug",
+    "django.core.context_processors.i18n",
+    "django.core.context_processors.media",
+    "django.core.context_processors.static",
+    )
+
+MAP_SHOW_MULTISELECT = False
+MAP_SHOW_DATE_RANGE = False
+
+GEOTIFF_DIR = os.path.join(BUILDOUT_DIR, 'var', 'geotiff')
+RADAR_NC_PATH = '/media/WORKSPACE/lizard-regenradar/radar.nc'
 
 try:
-    # Import local settings that aren't stored in svn/git.
+    # Import local settings that aren't stored in svn.
     from lizard_neerslagradar.local_testsettings import *
 except ImportError:
     pass