+++ lizard_neerslagradar/urls.py
@@ -1,9 +1,6 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-from django.conf.urls.defaults import include
-from django.conf.urls.defaults import patterns
-from django.conf.urls.defaults import url
+from django.conf.urls.defaults import *
+from django.conf import settings
 from django.contrib import admin
-from lizard_ui.urls import debugmode_urlpatterns
 
 from lizard_neerslagradar import views
 
@@ -11,14 +8,20 @@ admin.autodiscover()
 
 urlpatterns = patterns(
     '',
-    url(r'^ui/', include('lizard_ui.urls')),
-    # url(r'^map/', include('lizard_map.urls')),
-    url(r'^admin/', include(admin.site.urls)),
-    # url(r'^something/',
-    #     views.some_method,
-    #     name="name_it"),
-    # url(r'^something_else/$',
-    #     views.SomeClassBasedView.as_view(),
-    #     name='name_it_too'),
+    url(r'^$',
+        views.DefaultView.as_view(),
+        name='lizard_neerslagradar.default'),
+    url(r'^wms/$',
+        views.WmsView.as_view(),
+        name='lizard_neerslagradar.wms'),
+    (r'^map/', include('lizard_map.urls')),
     )
-urlpatterns += debugmode_urlpatterns()
+
+
+if settings.DEBUG:
+    # Add this also to the projects that use this application
+    urlpatterns += patterns(
+        '',
+        (r'', include('staticfiles.urls')),
+        (r'^admin/', include(admin.site.urls)),
+    )