+++ CHANGES.rst
@@ -6,3 +6,6 @@ Changelog of lizard-neerslagradar
 ----------------
 
 - Initial project structure created with nensskel 1.30.dev0.
+
+- Copied the code from Erik-Jan's prototype site, and got it to work
+  as a Lizard app.